sect11_215_14_refactoring_react_components_to_individual_javascript_modules
