+++ 13-full-stack/01-frontend/todo-app-myimp/src/components/todo/ErrorComponent.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+function ErrorComponent() {
+    return (
+        <div className="ErrorComponent">
+            <h1>We are working on the problem.</h1>
+            <div>
+                Apologies for the 404. Reach out to our team at ABC.
+            </div>
+        </div>
+    )
+}
+
+export default  ErrorComponent
