+++ 13-full-stack/01-frontend/todo-app-myimp/src/components/todo/FooterComponent.jsx
@@ -1,0 +1,11 @@
+function FooterComponent() {
+    return (
+        <footer className="footer">
+            <div className="container">
+                Your Footer
+            </div>
+        </footer>
+    )
+}
+
+export default FooterComponent
