+++ 13-full-stack/01-frontend/todo-app-myimp/src/components/todo/LoginComponent.jsx
@@ -1,0 +1,59 @@
+import React, {useState} from "react";
+import {useNavigate} from "react-router-dom";
+
+function LoginComponent() {
+    const [username, setUsername] = useState('omer');
+    const [password, setPassword] = useState('');
+    const [showSuccessMessage, setShowSuccessMessage] = useState(false);
+    const [showErrorMessage, setShowErrorMessage] = useState(false);
+    const navigate = useNavigate();
+    function handleUsernameChange(event) {
+        // console.log(event)
+        // console.log(event.target.value)
+        setUsername(event.target.value)
+    }
+
+    function handlePasswordChange(event) {
+        setPassword(event.target.value)
+    }
+
+    function handleSubmit() {
+        // console.log(username)
+        // console.log(password)
+
+        if(username === 'omer' && password === '123') {
+            console.log('Authentication is successful')
+            setShowSuccessMessage(true)
+            setShowErrorMessage(false)
+            navigate(`/welcome/${username}`)
+        } else {
+            console.log('Authentication is failed!')
+            setShowSuccessMessage(false)
+            setShowErrorMessage(true)
+        }
+    }
+
+    return (
+        <div className="Login">
+            <h1>Please Login.</h1>
+            {showSuccessMessage && <div className="successMessage">Authenticated Successfully</div>}
+            {showErrorMessage && <div className="errorMessage">Authentication Failed.
+                Please check your credentials.</div>}
+            <div className="LoginForm">
+                <div>
+                    <label>User Name</label>
+                    <input type="text" name="username" value={username} onChange={handleUsernameChange} />
+                </div>
+                <div>
+                    <label>Password</label>
+                    <input type="password" name="password" value={password} onChange={handlePasswordChange} />
+                </div>
+                <div>
+                    <button type="button" name="login" onClick={handleSubmit}>login</button>
+                </div>
+            </div>
+        </div>
+    )
+}
+
+export default LoginComponent
