+++ 13-full-stack/01-frontend/todo-app-myimp/src/components/todo/LogoutComponent.jsx
@@ -1,0 +1,12 @@
+function LogoutComponent() {
+    return (
+        <div className="LogoutComponent">
+            <h1>You're logged out!</h1>
+            <div>
+                Thank you for using our App. Come back soon!
+            </div>
+        </div>
+    )
+}
+
+export default LogoutComponent
