+++ 13-full-stack/01-frontend/todo-app-myimp/src/components/todo/TodoApp.jsx
@@ -1,6 +1,13 @@
 import './TodoApp.css'
-import {useState} from "react";
-import {BrowserRouter, Link, Route, Routes, useNavigate, useParams} from "react-router-dom";
+import React from "react";
+import {BrowserRouter, Route, Routes} from "react-router-dom";
+import LogoutComponent from "./LogoutComponent";
+import FooterComponent from "./FooterComponent";
+import HeaderComponent from "./HeaderComponent";
+import ListTodosComponent from "./ListTodosComponent";
+import ErrorComponent from "./ErrorComponent";
+import WelcomeComponent from "./WelcomeComponent";
+import LoginComponent from "./LoginComponent";
 
 export default function TodoApp() {
     return (
@@ -22,171 +29,3 @@ export default function TodoApp() {
     )
 }
 
-function LoginComponent() {
-    const [username, setUsername] = useState('omer');
-    const [password, setPassword] = useState('');
-    const [showSuccessMessage, setShowSuccessMessage] = useState(false);
-    const [showErrorMessage, setShowErrorMessage] = useState(false);
-    const navigate = useNavigate();
-    function handleUsernameChange(event) {
-        // console.log(event)
-        // console.log(event.target.value)
-        setUsername(event.target.value)
-    }
-
-    function handlePasswordChange(event) {
-        setPassword(event.target.value)
-    }
-
-    function handleSubmit() {
-        // console.log(username)
-        // console.log(password)
-
-        if(username === 'omer' && password === '123') {
-            console.log('Authentication is successful')
-            setShowSuccessMessage(true)
-            setShowErrorMessage(false)
-            navigate(`/welcome/${username}`)
-        } else {
-            console.log('Authentication is failed!')
-            setShowSuccessMessage(false)
-            setShowErrorMessage(true)
-        }
-    }
-
-    return (
-        <div className="Login">
-            <h1>Please Login.</h1>
-            {showSuccessMessage && <div className="successMessage">Authenticated Successfully</div>}
-            {showErrorMessage && <div className="errorMessage">Authentication Failed.
-                                                            Please check your credentials.</div>}
-            <div className="LoginForm">
-                <div>
-                    <label>User Name</label>
-                    <input type="text" name="username" value={username} onChange={handleUsernameChange} />
-                </div>
-                <div>
-                    <label>Password</label>
-                    <input type="password" name="password" value={password} onChange={handlePasswordChange} />
-                </div>
-                <div>
-                    <button type="button" name="login" onClick={handleSubmit}>login</button>
-                </div>
-            </div>
-        </div>
-    )
-}
-
-function WelcomeComponent() {
-    const {username} = useParams();
-
-    console.log(username)
-
-    return (
-        <div className="WelcomeComponent">
-            <h1>Welcome Omer</h1>
-            <div>
-                Manage your Todos - <Link to="/todos">Go here</Link>
-            </div>
-        </div>
-    )
-}
-
-function ErrorComponent() {
-    return (
-        <div className="ErrorComponent">
-            <h1>We are working on the problem.</h1>
-            <div>
-                Apologies for the 404. Reach out to our team at ABC.
-            </div>
-        </div>
-    )
-}
-
-function ListTodosComponent() {
-    const today = new Date()
-    const targetDate = new Date(today.getFullYear()+12, today.getMonth(), today.getDay())
-
-    const todos = [
-                                            {id: 1, description: 'Learn AWS', done: false, targetDate: targetDate},
-                                            {id: 2, description: 'Learn Full Stack Dev.', done: false, targetDate: targetDate},
-                                            {id: 3, description: 'Learn DevOps', done: false, targetDate: targetDate}
-                                        ]
-
-    return (
-        <div className="container">
-            <h1>Things You Want To Do!</h1>
-            <div>
-                <table className="table">
-                    <thead>
-                        <tr>
-                            <td>ID</td>
-                            <td>Description</td>
-                            <td>Is Done?</td>
-                            <td>Target Date</td>
-                        </tr>
-                    </thead>
-                    <tbody>
-                    {
-                        todos.map(
-                            todo => (
-                                <tr key={todo.id}>
-                                    <td>{todo.id}</td>
-                                    <td>{todo.description}</td>
-                                    <td>{todo.done.toString()}</td>
-                                    <td>{todo.targetDate.toDateString()}</td>
-                                </tr>
-                            )
-                        )
-                    }
-                    </tbody>
-                </table>
-            </div>
-        </div>
-    )
-}
-
-function HeaderComponent() {
-    return (
-        <header className="border-bottom border-light border-5 mb-5 p-2">
-            <div className="container">
-                <div className="row">
-                    <nav className="navbar navbar-expand-lg">
-                        <a className="navbar-brand ms-2 fs-2 fw-bold text-black" href="https://www.in28minutes.com">in28minutes</a>
-                        <div className="collapse navbar-collapse">
-                            <ul className="navbar-nav">
-                                <li className="nav-item fs-5"><Link className="nav-link" to="/welcome/in28minutes">Home</Link></li>
-                                <li className="nav-item fs-5"><Link className="nav-link" to="/todos">Todos</Link></li>
-                            </ul>
-                        </div>
-                        <ul className="navbar-nav">
-                            <li className="nav-item fs-5"><Link className="nav-link" to="/login">Login</Link></li>
-                            <li className="nav-item fs-5"><Link className="nav-link" to="/logout">Logout</Link></li>
-                        </ul>
-                    </nav>
-                </div>
-            </div>
-        </header>
-    )
-}
-
-function FooterComponent() {
-    return (
-        <footer className="footer">
-            <div className="container">
-                Your Footer
-            </div>
-        </footer>
-    )
-}
-
-function LogoutComponent() {
-    return (
-        <div className="LogoutComponent">
-            <h1>You're logged out!</h1>
-            <div>
-                Thank you for using our App. Come back soon!
-            </div>
-        </div>
-    )
-}
